+++ lib/src/context/deserialization_context_impl.dart
@@ -244,7 +244,7 @@ class DeserializationContextImpl extends DeserializationContext
     return requireCollection(
         subject,
         predicate,
-        (itemDeserializer) =>
+        ({itemDeserializer}) =>
             UnorderedItemsCollectorDeserializer(collector, itemDeserializer),
         itemDeserializer: deserializer);
   }
@@ -253,7 +253,8 @@ class DeserializationContextImpl extends DeserializationContext
   C requireCollection<C, T>(RdfSubject subject, RdfPredicate predicate,
       CollectionDeserializerFactory<C, T> collectionDeserializerFactory,
       {Deserializer<T>? itemDeserializer}) {
-    final deserializer = collectionDeserializerFactory(itemDeserializer);
+    final deserializer =
+        collectionDeserializerFactory(itemDeserializer: itemDeserializer);
 
     return require<C>(subject, predicate, deserializer: deserializer);
   }
@@ -262,7 +263,8 @@ class DeserializationContextImpl extends DeserializationContext
   C? optionalCollection<C, T>(RdfSubject subject, RdfPredicate predicate,
       CollectionDeserializerFactory<C, T> collectionDeserializerFactory,
       {Deserializer<T>? itemDeserializer}) {
-    final deserializer = collectionDeserializerFactory(itemDeserializer);
+    final deserializer =
+        collectionDeserializerFactory(itemDeserializer: itemDeserializer);
 
     return optional<C>(subject, predicate, deserializer: deserializer);
   }